Make shim reference opaque to webpack
ncc uses webpack to bundle.

Depending on how the path to the postprocessing shim is constructed,
webpack might think that it's a dynamic import and attempt to bundle
the shim into the compiled output. This will make webpack/ncc error,
because the shim is a deno script.

What's worse, because `npm run dist` is a two-phase operation,
it's possible to make a change and not notice that it broke things
if you're only re-running the ncc step (`npm run package`).

Computers were a mistake, change my mind

Add deno cwd logging

More debugging of deno relative import path

More debugging

Maybe now?